# Use minified and optimized React for SSR