--- a/starter/files/package.json
+++ b/starter/files/package.json
@@ -24,6 +24,7 @@
     "babel-preset-env": "^1.3.2",
     "babel-preset-react": "^6.23.0",
     "babel-register": "^6.24.0",
+    "babili-webpack-plugin": "^0.0.11",
     "chalk": "^1.1.3",
     "compression-webpack-plugin": "^0.4.0",
     "css-loader": "^0.28.0",