… starter/files/yarn.lock