Sasha Aickin has a great [talk](https://www.youtube.com/watch?v=PnpfGy7q96U) where he details steps to improve the performance of SSR React apps. As I was getting ~300ms response times in production. I implemented setting production mode using the Define Plugin, and whitelisted `react` and `react-dom/server` from the externals plugin. I also added a minifier to remove all the dead code from React as well. I couldn't use Uglify because of all the ES2015+ plus syntax on the server so instead I used Babili.

The result was still ~300ms response times. I then realized that the response time from the graphQL server was likely to blame. Turns out the graph.cool server is in Ireland which happens to be pretty far from me. Knowing that ideally the API and page server would be located in the same data-center and the time to fetch data would be minimal, I temporarily commented out `getDataFromTree`. In the original version I was getting around 20 ms for a page response, while with the optimizations I was getting around 10 ms.

At that level the difference is imperceptible, but with a bigger app with more dom nodes, I would expect that this difference would increase as demonstrated in Sasha's talk. The downside to this change here is Babili on the server which is slower than using Uglify, and far slower than using nothing before. On my computer this took the server build from 10s to 25s. As an alternative, the pre-minified version of React and React-DOM can be used instead which is what Sasha did in his demo. I think the best way to do that is using a webpack resolve alias. Another approach to speed up the build would be to use concurrently instead of using webpack's multi compiler mode so that multiple cores can be used. This has the result of bringing the build time back down to where it began.

So first off, I might want to test a DOM node heavy page to see if these changes are still worth it. Also I have doubts on how I changed the server build process. Would using the pre-minified React be a better than this approach? Would it also be worth implementing concurrently in this PR or is that too far out of scope?