@leebenson 2017-04-16
Interesting consideration.

I actually tried Babili before Uglify in a pre-release version, and wound up yanking it out for the same reason - significant build times. Secondary to that was the build size was generally larger than Uglify (only a concern in the browser, of course.)

The 300ms response time was definitely 90-95% comprised of the example graph.cool call. Keeping the `getDataFromTree` call but swapping the `GraphQLMessage` component to this 'plain' version, sans-GraphQL query:

```js
const GraphQLMessage = () => (
  <Message data={{ allMessages: ['Test without GraphQL'] }} />
);
```

... yields consistent response times of **4-15ms** on my local machine with the current Uglify config:

<img width="308" alt="screen shot 2017-04-16 at 11 26 29" src="https://cloud.githubusercontent.com/assets/719722/25070482/9d0242da-2297-11e7-8938-d9d95cafa1d6.png">

Response times circa this range are problematic to gauge-- hitting 'refresh' in the browser can yield 1-4x the original timing, just based on whatever the CPU happens to be doing at the time. Concurrent refreshes about half a second apart yielded 9.209ms, 12.953ms, 4.658ms and 14.273ms... **the highest being 3.06x slower than the fastest.**

Exact same query, no external calls.

So, my feeling at this stage is that enduring an extra 15 second build time is overkill. I like the idea of boosting performance by using a pre-minified build, though - that seems like a freebie without any downside. In the talk you referenced, this code given:

<img width="800" alt="screen shot 2017-04-16 at 11 14 51" src="https://cloud.githubusercontent.com/assets/719722/25070512/71427498-2298-11e7-9040-a0a983e1f4fd.png">

Like you said, that'd just be a case of throwing in a resolve alias in Webpack.

What do you think? Want to test the above?