@leebenson 2017-04-16
Scratch the Webpack resolve alias, it should actually just be a case of making the switch in `kit/entry/server.js`