@leebenson 2017-04-16
I've added the production `process.env.NODE_ENV` var to the [kit repo](https://github.com/reactql/kit), to remove React debugging with the SSR version.

I tried switching `react-dom/server` -> `react/dist/react.min`, but it looks like that's no longer valid because the `renderToString` method is missing. There are a few `dist` versions inside react-dom, but they dump their own errors when trying to use.

@jtmthf, if you want to test/work on this further, please open a PR at the [kit repo](https://github.com/reactql/kit). I'll close this one for now, since the CLI has now been relegated to a pure build tool in [2.2.0](https://github.com/reactql/cli/releases/tag/2.2.0)+